pyproject.toml
[project]
name = "duckdb-pipe"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Evan-Kim2028", email = "ekcopersonal@gmail.com" }
]
dependencies = [
    "ipykernel>=6.29.5",
    "duckdb>=1.1.2",
    "hypermanager>=0.1.9",
    "dlt>=1.3.0",
    "pyarrow>=18.0.0",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/duckdb_pipe"]
